+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.6:
 
+* Remove "appcache_ignorable_permutation_properties" configuration property
+  and use builtin mechanisms for determining selection properties.
 * Generate a more complete Maven POM file.
 * Compile against the last release of GWT 2.5.1 rather than the 2.6 RC1.
 * Add javadoc packages to the build process.